+++ src/components/Expenses/ExpenseItem.css
@@ -29,7 +29,7 @@
     font-size: 1rem;
     font-weight: bold;
     color: white;
-    background-color: #40005d;
+    background-color: #005d55;
     border: 1px solid white;
     padding: 0.5rem;
     border-radius: 12px;
