+++ src/components/Expenses/Expenses.css
@@ -1,6 +1,6 @@
 .expenses {
     padding: 1rem;
-    background-color: rgb(31,31,31);
+    background-color: rgb(1, 1, 1);
     margin: 2rem, auto;
     width: 50rem;
     max-width: 95%;
